Clean up namespace methods
The `_add_named_storage()` method contained code that was otherwise
duplicated, but thee code didn't actually match the type arguments
of the class it resided in.

One solution would have been to insert an intermediate class for
housing `_add_named_storage()`, but as its only 4 lines of code,
I think that wouldn't actually make the code easier to maintain.