Improve truncation and sign extension of fixed values
Previously, `FixedValue` asserted in its constructor that the given
expression mask fit within the given width. There were several
problems with this:
- `AssertionError` is not the right exception type for the situation
- for signed values the mask cannot be used in this way
- the computed mask is not guaranteed to be as tight as possible

I decided to mimic the behavior of the instruction set definition
language and truncate the value expression in `FixedValue`. This
does mean that mismatches between expressoin and bit string width
will be harder to spot, but compared to requiring the callers to
add trunctation, that seems worth it.

I also added sign extension to `FixedValueReference.expr`, to make it
consistent with how `Reference` loads signed values. This obsoleted
the sign extension applied in the asm formatting of negative integers.

These added trunctations and sign extensions broke detection of
`ImmediateValue` placeholders. To solve that, I added a special case
in the sign extension simplifier and made `FixedValueReference`
simplify its expression.