Create placeholder dictionary instead of namespace for decoding
Inheritance wasn't being used here as the `elements` items were
accessed rather than the namespace itself.

Some location information is lost, although in practical terms
that was already lost in the previous commit, as the reconstructed
namespace definitions received `None` for the location. It could
be brought back either in `ModeEntry` or a `FixedValueReference`
subclass, but I prefer to delay that until the final shape of the
code is more clear.