Dismantle combined storage of instructions and mode entries in parser
The mode entries are now available in the same form in the `modes`
dictionary, so they are redundant in the combined storage. And with
the modes removed, there was only a single key (`None`) left for the
instructions, so the dictionary was redundant as well.

The old combined dictionary is still constructed just before the
creation of the `InstructionSet` instance, but that should be
temporary.