Get rid of `ParsedModeEntry`
The extra data it held was moved into `ModeEntry` itself. I'm unsure
whether it belongs there: maybe it should be in `Encoding` instead.

The `create_mode_entry()` function is now in the `decode` module
to avoid circular imports, but that's not the place it should land
in the end.

So all in all, this commit is pretty rough, but I think it's moving
in the right direction of validating mode entry consistency for each
entry individually.