Compute decoding in `Encoding` factory method
The decoding is a reversal of the encoding, so it makes sense to keep
them together.

However, I think this commit actually goes a bit too far, by also
putting the check whether the entire context can be decoded into
the `Encoding` factory, which seems like something that should be
enforced at the mode entry level instead.