Improve sign extension unit tests
Removed tests of the form `X = X`, as they will always pass, since we
simplify both sides. I tried an alternative where only the left hand
side is simplified, but it turns out it's not easy to predict what
an expression simplifies to.

Converted a test case that was still checking expressions in Python
to the docstring format.