Rename `Negation` to `ZeroTest`
This is a more direct description of what the operation does: there is
no confusing it with a (bitwise or intger) complement. Also we already
have a separation between high-level operators and low-level ones,
for example `SignTest` also only exists at the low level.

The unit test deliberately still use "negation", as those test high
level expressions (definition language) and not the low-level ones
(`Expression` objects).