Take mode out of `EncodingMultiMatch`
By storing the attributes derived from the mode instead of the mode
itself, we will be able to apply restrictions on them, like whether
they can be `None` or not.

From a design perspective, the encoding definition doesn't need to
know what it is encoding *for*, just the shape of the encoding.
The context is where the connection to the submode is made.