Give `EncodingMultiMatch` a consistent width
Previously, every `X@` placeholder was parsed 1:1 into an
`EncodingMultiMatch`, which meant that class had to deal with
a lot of different scenarios and therefore became quite complex.

This commit lets the parser handle two particular scenarios so
`EncodingMultiMatch` no longer has to:
- empty encodings
- multi-part encodings where the first (primary) item has a different
  width from the following (auxiliary) items

Put together, this means `EncodingMultiMatch` only needs a single
width property and that width cannot be `None`.