Remove the notion of encoding type
Both the encoding items and the instruction set as a whole had
a property to get its encoding type, which was always an unsigned
integer type matching the encoding width. This property was used
only in one place, to get the width from the type, while that
width could have been fetched directly.

From more of a design standpoint, our encoding is a series of bit
strings, which don't have a type: our bit strings were designed to
only have a width (number of bits) and the typing is applied by
the reference wrapping the bit string. Dropping the type from
the encoding is consistent with that.